requirements.txt
# TebeoSfera Scraper - Requirements
# Python 3.6+

# Required for GUI
Pillow>=8.0.0
rarfile>=4.0
